% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_lfsr.R
\name{compute_marginal_mean_var}
\alias{compute_marginal_mean_var}
\title{Output all the marginal mean and variance for each PSD value}
\usage{
compute_marginal_mean_var(
  data_i,
  psd_values,
  refined_x,
  Si = NULL,
  Omegai = NULL,
  num_basis = 30,
  betaprec = 1e-06,
  order = 1,
  pred_step = 1,
  likelihood,
  deriv = 0
)
}
\description{
Output all the marginal mean and variance for each PSD value
}
\keyword{internal}
