% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_likelihood.R
\name{fash_L_compute}
\alias{fash_L_compute}
\title{Compute the L Matrix for a Set of Datasets}
\usage{
fash_L_compute(
  fash_data,
  likelihood = "gaussian",
  num_cores = 1,
  grid = seq(0, 2, length.out = 10),
  pred_step = 1,
  num_basis = 30,
  betaprec = 1e-06,
  order = 1,
  verbose = FALSE
)
}
\arguments{
\item{fash_data}{The output from \code{fash_set_data}, containing preprocessed datasets.}

\item{likelihood}{A character string specifying the likelihood function to use. Options are `gaussian` and `poisson`.}

\item{num_cores}{An integer specifying the number of cores to use for parallel processing.}

\item{grid}{A numeric vector representing the grid of PSD (Predictive Standard Deviation) values.}

\item{pred_step}{A numeric value specifying the prediction step size.}

\item{num_basis}{An integer specifying the number of O-Spline basis functions to use for the approximation.}

\item{betaprec}{A numeric value representing the precision of the fixed effects coefficients.}

\item{order}{An integer specifying the order of the Integrated Wiener Process (IWP) prior.}

\item{verbose}{A logical value. If \code{TRUE}, shows a progress bar when \code{num_cores = 1}.}
}
\value{
A numeric matrix where each row corresponds to a dataset and each column corresponds to a grid value of PSD.
}
\description{
Computes the L matrix, where each row corresponds to a dataset and each column corresponds to a grid value of PSD (Predictive Standard Deviation).
Handles both Gaussian and Poisson likelihoods using helper functions.
}
\examples{
# Example usage
set.seed(1)
Y <- matrix(rpois(20, lambda = 5), nrow = 4, ncol = 5)
S <- matrix(1, nrow = 4, ncol = 5)
smooth_var <- matrix(runif(20), nrow = 4, ncol = 5)
offset <- 1
fash_data_poisson <- fash_set_data(Y = Y, smooth_var = smooth_var, offset = offset)
fash_data_gaussian <- fash_set_data(Y = Y, smooth_var = smooth_var, S = S)
grid <- seq(0.1, 2, length.out = 10)
L_matrix_gaussian <-
  fashr:::fash_L_compute(fash_data_gaussian, likelihood = "gaussian",
                         num_cores = 1, grid = grid, num_basis = 30,
                         betaprec = 1e-6, order = 2, verbose = TRUE)
L_matrix_poisson <-
  fashr:::fash_L_compute(fash_data_poisson, likelihood = "poisson",
                         num_cores = 1, grid = grid, num_basis = 30,
                         betaprec = 1e-6, order = 2, verbose = TRUE)

}
\keyword{internal}
